Phase 2: 型安全性の確保 - Type safety improvements
- Fixed Object  String conversion in referral_service.dart
- Added public logEvent method to AnalyticsService
- Fixed num  double conversion in quest_recommendation_service.dart
- Fixed ProofType enum comparison in today_logs_screen.dart
- Fixed Padding children parameter in changelog_screen.dart
- Fixed MinqTheme import and property access in quest_attributes_selector.dart
- Added networkStatusProvider to network_status_service.dart
- Updated offline_mode_indicator.dart to handle AsyncValue from StreamProvider

+++ lib/domain/recommendation/quest_recommendation_service.dart
@@ -43,7 +43,7 @@ class QuestRecommendationService {
       final recencyBoost = _recencyWeight(completionsByQuest[quest.id]?.lastCompletedAt, now);
       final baseScore = (1 - completionRate) * 0.6 + recencyBoost + streak * 0.05;
       final variance = _categoryDiversityWeight(quest.category, categoryCounts);
-      final totalScore = max(0, baseScore + variance);
+      final totalScore = max(0.0, baseScore + variance).toDouble();
       final reason = _buildReason(
         quest: quest,
         completionRate: completionRate,
@@ -119,8 +119,8 @@ class QuestRecommendationService {
     }
     final occurrences = categoryCounts[category] ?? 0;
     if (occurrences == 0) return 0.2;
-    final average = categoryCounts.values.fold<int>(0, (prev, value) => prev + value) /
-        categoryCounts.values.length;
+    final average = (categoryCounts.values.fold<int>(0, (prev, value) => prev + value) /
+        categoryCounts.values.length).toDouble();
     if (occurrences > average) {
       return -0.1;
     }
